+++ Kite-SDK/PSPrintSDK/OLAnalytics.m
@@ -56,9 +56,8 @@ static NSString *nonNilStr(NSString *str) {
 }
 
 @interface OLProduct (Private)
-
 - (NSDecimalNumber*) unitCostDecimalNumber;
-
+- (NSString *)currencyCode;
 @end
 
 @interface OLKitePrintSDK (Private)
@@ -266,6 +265,7 @@ static NSString *nonNilStr(NSString *str) {
     OLProductTemplate *template = [OLProductTemplate templateWithId:[job templateId]];
     if (template){
         dict[kOLAnalyticsProductName] = template.name;
+        dict[kOLAnalyticsProductId] = template.identifier;
         dict[kOLAnalyticsNumberOfPhotosInItem] = [NSNumber numberWithInteger:[job quantity]];
         dict[kOLAnalyticsQuantity] = [NSNumber numberWithInteger:[job extraCopies]+1];
         
@@ -300,6 +300,7 @@ static NSString *nonNilStr(NSString *str) {
             dict[kOLAnalyticsNumberOfPhotosInOrder] = [NSNumber numberWithInteger:order.totalAssetsToUpload];
             dict[kOLAnalyticsPromoCode] = order.promoCode;
             dict[kOLAnalyticsCurrencyCode] = order.currencyCode;
+            dict[kOLAnalyticsOrderId] = order.receipt;
             [order costWithCompletionHandler:^(OLPrintOrderCost *cost, NSError *error){
                 dict[kOLAnalyticsOrderCost] = [cost totalCostInCurrency:order.currencyCode];
                 dict[kOLAnalyticsOrderShippingCost] = [cost shippingCostInCurrency:order.currencyCode];
@@ -354,14 +355,15 @@ static NSString *nonNilStr(NSString *str) {
     [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsEventLevel : @1}];
 }
 
-+ (void)trackProductDetailsScreenViewed:(NSString *)productName hidePrice:(BOOL)hidePrice{
++ (void)trackProductDetailsScreenViewed:(OLProductTemplate *)productTemplate hidePrice:(BOOL)hidePrice{
     NSString *eventName = kOLAnalyticsEventNameProductDetailsScreenViewed;
     NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
-    [dict[@"properties"] setObject:nonNilStr(productName) forKey:@"Product Name"];
+    [dict[@"properties"] setObject:nonNilStr(productTemplate.name) forKey:kOLAnalyticsProductName];
+    [dict[@"properties"] setObject:nonNilStr(productTemplate.identifier) forKey:kOLAnalyticsProductId];
     [dict[@"properties"] setObject:hidePrice ? @"YES" : @"NO" forKey:@"Hide Price on Product Description"];
     [OLAnalytics sendToMixPanelWithDictionary:dict];
     
-    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : nonNilStr(productName), kOLAnalyticsEventLevel : @1}];
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : nonNilStr(productTemplate.name), kOLAnalyticsProductId : nonNilStr(productTemplate.identifier), kOLAnalyticsEventLevel : @1}];
 }
 
 + (void)trackProductListScreenViewedWithTemplateClass:(NSString *)templateClassString{
@@ -547,6 +549,32 @@ static NSString *nonNilStr(NSString *str) {
     [OLAnalytics sendToMixPanelWithDictionary:dict];
     
     [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:printOrder extraInfo:@{kOLAnalyticsApplePayAvailable : applePayIsAvailable, kOLAnalyticsPaymentMethod : methodName,kOLAnalyticsEventLevel : @1}];
+}
+
++ (void)trackItemAddedToBasket:(id<OLPrintJob>)item{
+    NSString *eventName = kOLAnalyticsEventNameItemAddedToBasket;
+    
+    OLProduct *product = [OLProduct productWithTemplateId:item.templateId];
+    OLProductTemplate *template = [OLProductTemplate templateWithId:item.templateId];
+    NSString *currency = product.currencyCode;
+    NSDecimalNumber *price = [product unitCostDecimalNumber];
+    NSString *itemType = template.templateClass;
+    
+    if (!template || !product){
+        return;
+    }
+
+    NSDictionary *dict = [OLAnalytics defaultDictionaryForEventName:eventName];
+    [dict[@"properties"] setObject:template.name forKey:kOLAnalyticsProductName];
+    [dict[@"properties"] setObject:template.identifier forKey:kOLAnalyticsProductId];
+    [dict[@"properties"] setObject:price forKey:kOLAnalyticsItemPrice];
+    [dict[@"properties"] setObject:currency forKey:kOLAnalyticsCurrencyCode];
+    if (itemType){
+        [dict[@"properties"] setObject:itemType forKey:kOLAnalyticsProductCategory];
+    }
+    [OLAnalytics sendToMixPanelWithDictionary:dict];
+
+    [OLAnalytics reportAnalyticsEventToDelegate:eventName job:nil printOrder:nil extraInfo:@{kOLAnalyticsProductName : template.name, kOLAnalyticsProductId : template.identifier, kOLAnalyticsItemPrice : price, kOLAnalyticsCurrencyCode : currency, kOLAnalyticsProductName: nonNilStr(itemType), kOLAnalyticsEventLevel : @1}];
 }
 
 #pragma mark Track Secondary Events - Not Sent to MixPanel