+++ Kite-SDK/PSPrintSDK/OLCaseViewController.m
@@ -412,6 +412,8 @@
         return;
     }
     
+    BOOL fromEdit = NO;
+    
     OLPrintOrder *printOrder = [OLUserSession currentSession].printOrder;
     OLProductPrintJob *job = [[OLProductPrintJob alloc] initWithTemplateId:self.product.templateId OLAssets:@[asset]];
     for (NSString *option in self.product.selectedOptions.allKeys){
@@ -423,12 +425,18 @@
             job.dateAddedToBasket = [existingJob dateAddedToBasket];
             job.extraCopies = existingJob.extraCopies;
             [printOrder removePrintJob:existingJob];
+            fromEdit = YES;
         }
     }
     [job.acceptedOffers addObjectsFromArray:self.product.acceptedOffers.allObjects];
     [job.declinedOffers addObjectsFromArray:self.product.declinedOffers.allObjects];
     job.redeemedOffer = self.product.redeemedOffer;
     [printOrder addPrintJob:job];
+#ifndef OL_NO_ANALYTICS
+    if (!fromEdit){
+        [OLAnalytics trackItemAddedToBasket:job];
+    }
+#endif
     
     [printOrder saveOrder];
     
@@ -492,6 +500,7 @@
 
 - (void)applyDownloadedMask {
     if (self.downloadedMask){
+        [self maskWithImage:self.maskImage targetView:self.artboard];
         return;
     }
     
@@ -755,7 +764,7 @@
 }
 
 - (void)renderImageWithCompletionHandler:(void (^)(void))handler{
-    if (![self isUsingMultiplyBlend]  || self.maskActivityIndicator.isAnimating || [[[UIDevice currentDevice] systemVersion] floatValue] < 10){
+    if (![self isUsingMultiplyBlend]  || self.maskActivityIndicator.isAnimating || [[[UIDevice currentDevice] systemVersion] floatValue] < 10 || self.presentedViewController){
         if (handler){
             handler();
         }