+++ Kite-SDK/PSPrintSDK/OLKiteViewController.m
@@ -137,12 +137,12 @@ static CGFloat fadeTime = 0.3;
 
 - (instancetype _Nullable)initWithAssets:(NSArray <OLAsset *>*_Nonnull)assets info:(NSDictionary *_Nullable)info{
     [OLAnalytics setExtraInfo:info];
-    if ((self = [[UIStoryboard storyboardWithName:@"OLKiteStoryboard" bundle:[OLKiteUtils kiteResourcesBundle]] instantiateViewControllerWithIdentifier:@"KiteViewController"])) {
-        [OLUserSession currentSession].appAssets = assets;
-        [[OLUserSession currentSession] resetUserSelectedPhotos];
-        [OLUserSession currentSession].printOrder.userData = info;
-    }
+    NSArray <OLAsset *>*assetsCopy = [assets copy]; // Prevents assets being nilled in some cases
+    self = [[UIStoryboard storyboardWithName:@"OLKiteStoryboard" bundle:[OLKiteUtils kiteResourcesBundle]] instantiateViewControllerWithIdentifier:@"KiteViewController"];
     [OLKiteABTesting sharedInstance].launchedWithPrintOrder = NO;
+    [OLUserSession currentSession].appAssets = assetsCopy;
+    [[OLUserSession currentSession] resetUserSelectedPhotos];
+    [OLUserSession currentSession].printOrder.userData = info;
     
     return self;
 }
@@ -248,7 +248,9 @@ static CGFloat fadeTime = 0.3;
         [[self.view viewWithTag:9999] removeFromSuperview];
     }
     
-    self.view.backgroundColor = [self.loadingImageView.image colorAtPixel:CGPointMake(3, 3)];
+    if (self.loadingImageView){
+        self.view.backgroundColor = [self.loadingImageView.image colorAtPixel:CGPointMake(3, 3)];
+    }
     
     [self transitionToNextScreen];
 }
@@ -333,7 +335,7 @@ static CGFloat fadeTime = 0.3;
         UIStoryboard *sb = [UIStoryboard storyboardWithName:@"OLKiteStoryboard" bundle:[OLKiteUtils kiteResourcesBundle]];
         UIViewController *vc;
         OLProduct *product;
-        if (groups.count == 0 && !([OLProductTemplate templates].count != 0 && [OLKiteABTesting sharedInstance].launchedWithPrintOrder)) {
+        if (groups.count == 0 && ![OLProductTemplate isSyncInProgress] && !([OLProductTemplate templates].count != 0 && [OLKiteABTesting sharedInstance].launchedWithPrintOrder)) {
                 UIAlertController *ac = [UIAlertController alertControllerWithTitle:NSLocalizedStringFromTableInBundle(@"Store Maintenance", @"KitePrintSDK", [OLKiteUtils kiteLocalizationBundle], @"") message:NSLocalizedStringFromTableInBundle(@"Our store is currently undergoing maintenance so no products are available for purchase at this time. Please try again a little later.", @"KitePrintSDK", [OLKiteUtils kiteLocalizationBundle], @"") preferredStyle:UIAlertControllerStyleAlert];
                 [ac addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTableInBundle(@"OK", @"KitePrintSDK", [OLKiteUtils kiteLocalizationBundle], @"Acknowledgent to an alert dialog.") style:UIAlertActionStyleDefault handler:^(UIAlertAction *action) {
                     [welf dismiss];