+++ Kite-SDK/CHANGELOG.md
@@ -1,3 +1,23 @@
+### 6.8.10
+- Fixed: Launching the SDK with some assets would sometimes result in those assets being ignored.
+
+### 6.8.9
+- Fixed: Custom image pickers not getting the maximum amount of photos and the currently selected assets
+
+### 6.8.8
+- Fixed: An issue with remote configuration
+
+### 6.8.6
+- Changed: Adding an item with more images than the pack size to the basket will not split it into 2 items in the basket.
+- Fixed: Custom image pickers not getting the maximum amount of photos for some products
+- Fixed: Minor layout issues on iPhone X
+- Added: Order ID to the relevant analytics event
+- Added: Add to basket analytics event
+- Updated: ES, NL localizations
+
+### 6.8.5
+- Fixed: The case where the user could go back from the basket under some circumstances.
+
 ### 6.8.4
 - Changed: The estimated delivery times are now clarified as 'working days'.